requirements.txt
#!/usr/bin/env python3
# Requirements for setup.py script
# Install with: pip install -r requirements.txt

web3>=6.0.0
eth-account>=0.11.0
python-dotenv>=1.0.0
